Tidy up things, refactoring
