+++ src/main/java/homewatch/net/HttpCachingUtils.java
@@ -4,13 +4,10 @@ import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
 import homewatch.exceptions.NetworkException;
-import homewatch.net.HttpUtils;
-import homewatch.net.ThingResponse;
 import okhttp3.HttpUrl;
 
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
-import java.util.logging.Logger;
 
 public class HttpCachingUtils {
   private static final LoadingCache<HttpUrl, ThingResponse> cachedResponses = CacheBuilder.newBuilder()
