+++ src/main/java/homewatch/server/ClassDiscoverer.java
@@ -6,6 +6,9 @@ import org.reflections.Reflections;
 import java.util.Set;
 
 class ClassDiscoverer {
+  private ClassDiscoverer() {
+  }
+
   public static Set<Class<? extends Thing>> getThings() {
     Reflections reflections = new Reflections("homewatch.things");
     return reflections.getSubTypesOf(Thing.class);
