+++ src/main/java/homewatch/things/Thing.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 
 public interface Thing {
   @JsonIgnore
-  public ThingServiceFactory<? extends Thing> getFactory();
+  ThingServiceFactory getFactory();
 
   @JsonIgnore
-  public String getStringRepresentation();
+  String getStringRepresentation();
 }
