+++ src/test/java/homewatch/server/TestNgrok.java
@@ -2,10 +2,10 @@ package homewatch.server;
 
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import homewatch.constants.JsonUtils;
-import homewatch.stubs.NgrokStubs;
 import homewatch.exceptions.NetworkException;
 import homewatch.net.HttpUtils;
 import homewatch.net.ThingResponse;
+import homewatch.stubs.NgrokStubs;
 import homewatch.things.ServerRunner;
 import okhttp3.HttpUrl;
 import org.junit.Rule;
