+++ src/test/java/homewatch/things/services/lights/TestHueLight.java
@@ -1,16 +1,13 @@
 package homewatch.things.services.lights;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
-import homewatch.constants.JsonUtils;
-import homewatch.stubs.LightStubs;
 import homewatch.exceptions.NetworkException;
+import homewatch.stubs.LightStubs;
 import homewatch.things.ThingService;
 import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 
-import java.io.File;
 import java.io.IOException;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
@@ -32,8 +29,6 @@ public class TestHueLight {
   public static void setup() throws IOException {
     lightService = new HueLightService(InetAddress.getLocalHost().getHostName(), PORT);
     lightService.setLightID(1);
-
-    JsonNode HUE_RESPONSE = JsonUtils.getOM().readTree(new File("src/test/fixtures/hue/get.json"));
   }
 
   @Test
