+++ src/test/java/homewatch/things/services/lights/TestRestLight.java
@@ -1,8 +1,8 @@
 package homewatch.things.services.lights;
 
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
-import homewatch.stubs.LightStubs;
 import homewatch.exceptions.NetworkException;
+import homewatch.stubs.LightStubs;
 import homewatch.things.ThingService;
 import org.junit.BeforeClass;
 import org.junit.Rule;
