+++ src/test/java/homewatch/things/services/locks/TestLockController.java
@@ -3,8 +3,8 @@ package homewatch.things.services.locks;
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.http.exceptions.UnirestException;
-import homewatch.stubs.LockStubs;
 import homewatch.exceptions.NetworkException;
+import homewatch.stubs.LockStubs;
 import homewatch.things.ServerRunner;
 import org.json.JSONObject;
 import org.junit.BeforeClass;
