+++ src/test/java/homewatch/things/services/motionsensors/TestMotionSensorController.java
@@ -4,8 +4,8 @@ import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.http.exceptions.UnirestException;
 import homewatch.constants.JsonUtils;
-import homewatch.stubs.MotionSensorStubs;
 import homewatch.exceptions.NetworkException;
+import homewatch.stubs.MotionSensorStubs;
 import homewatch.things.ServerRunner;
 import org.junit.Before;
 import org.junit.Rule;
