+++ src/test/java/homewatch/things/services/thermostat/TestThermostatController.java
@@ -5,8 +5,8 @@ import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.http.exceptions.UnirestException;
 import homewatch.constants.JsonUtils;
-import homewatch.stubs.ThermostatStubs;
 import homewatch.exceptions.NetworkException;
+import homewatch.stubs.ThermostatStubs;
 import homewatch.things.ServerRunner;
 import org.json.JSONObject;
 import org.junit.BeforeClass;
