+++ src/test/java/homewatch/things/services/weather/TestWeatherController.java
@@ -4,8 +4,8 @@ import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.http.exceptions.UnirestException;
 import homewatch.constants.JsonUtils;
-import homewatch.stubs.WeatherStubs;
 import homewatch.exceptions.NetworkException;
+import homewatch.stubs.WeatherStubs;
 import homewatch.things.ServerRunner;
 import org.junit.Before;
 import org.junit.Rule;
